# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyPay fintech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: List user deposits
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositsResult"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositCreated"

  # ── WITHDRAWALS ──────────────────────────────────────────────
...
components:
  schemas:
...
    DepositItem:
      type: object
      properties:
        id:
          type: string
        amount:
          type: string
        utrNumber:
          type: string
        paymentMethod:
          type: string
        screenshotUrl:
          type: ["string", "null"]
        status:
          type: string
        adminNotes:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    DepositsResult:
      type: object
      properties:
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/DepositItem"
        hasMore:
          type: boolean
    DepositInput:
      type: object
      required: [amount, utrNumber, paymentMethod]
      properties:
        amount:
          type: number
        utrNumber:
          type: string
        paymentMethod:
          type: string
        screenshotUrl:
          type: string
    DepositCreated:
      type: object
      properties:
        success:
          type: boolean
        deposit:
          $ref: "#/components/schemas/DepositItem"

    # ── Withdrawals ──────────────────────────────────────────
...